add color to dbt output
