+++ dbt/runner.py
@@ -1,5 +1,3 @@
-from __future__ import print_function
-
 import hashlib
 import psycopg2
 import os
@@ -19,19 +17,13 @@ import dbt.tracking
 import dbt.schema
 import dbt.graph.selector
 import dbt.model
+import dbt.printer as printer
 
 from multiprocessing.dummy import Pool as ThreadPool
 
 
-from colorama import Fore, Back, Style
-
-
 ABORTED_TRANSACTION_STRING = ("current transaction is aborted, commands "
                               "ignored until end of transaction block")
-
-
-def get_timestamp():
-    return time.strftime("%H:%M:%S")
 
 
 def get_hash(model):
@@ -46,151 +38,44 @@ def is_enabled(model):
     return model.get('config', {}).get('enabled') is True
 
 
-def print_timestamped_line(msg):
-    logger.info("{} | {}".format(get_timestamp(), msg))
-
-
-def print_fancy_output_line(msg, status, index, total, execution_time=None):
-    prefix = "{timestamp} | {index} of {total} {message}".format(
-        timestamp=get_timestamp(),
-        index=index,
-        total=total,
-        message=msg)
-    justified = prefix.ljust(80, ".")
-
-    if execution_time is None:
-        status_time = ""
-    else:
-        status_time = " in {execution_time:0.2f}s".format(
-            execution_time=execution_time)
-
-    colors = {"ERROR" : Fore.RED, "SKIP" : Fore.YELLOW}
-    status_txt = colors.get(status, Fore.GREEN) + status + Style.RESET_ALL
-    output = "{justified} [{status}{status_time}]".format(
-        justified=justified, status=status_txt, status_time=status_time)
-
-    logger.info(output)
-
-
-def print_skip_line(model, schema, relation, index, num_models):
-    msg = 'SKIP relation {}.{}'.format(schema, relation)
-    skip_txt = Fore.YELLOW + 'SKIP' + Style.RESET_ALL
-    print_fancy_output_line(msg, skip_txt, index, num_models)
-
-
-def print_counts(flat_nodes):
-    counts = {}
-
-    for node in flat_nodes:
-        t = node.get('resource_type')
-
-        if node.get('resource_type') == NodeType.Model:
-            t = '{} {}'.format(get_materialization(node), t)
-
-        counts[t] = counts.get(t, 0) + 1
-
-    stat_line = ", ".join(
-        ["{} {}s".format(v, k) for k, v in counts.items()])
-
-    logger.info("")
-    print_timestamped_line("Running {}".format(stat_line))
-    print_timestamped_line("")
-
-
 def print_start_line(node, schema_name, index, total):
     if is_type(node, NodeType.Model):
-        print_model_start_line(node, schema_name, index, total)
+        printer.print_model_start_line(node, schema_name, index, total)
     if is_type(node, NodeType.Test):
-        print_test_start_line(node, schema_name, index, total)
+        printer.print_test_start_line(node, schema_name, index, total)
     if is_type(node, NodeType.Archive):
-        print_archive_start_line(node, index, total)
-
-
-def print_test_start_line(model, schema_name, index, total):
-    msg = "START test {name}".format(
-        name=model.get('name'))
-
-    print_fancy_output_line(msg, 'RUN', index, total)
-
-
-def print_model_start_line(model, schema_name, index, total):
-    msg = "START {model_type} model {schema}.{relation}".format(
-        model_type=get_materialization(model),
-        schema=schema_name,
-        relation=model.get('name'))
-
-    print_fancy_output_line(msg, 'RUN', index, total)
-
-
-def print_archive_start_line(model, index, total):
-    cfg = model.get('config', {})
-    msg = "START archive {source_schema}.{source_table} --> "\
-          "{target_schema}.{target_table}".format(**cfg)
-
-    print_fancy_output_line(msg, 'RUN', index, total)
+        printer.print_archive_start_line(node, index, total)
 
 
 def print_result_line(result, schema_name, index, total):
     node = result.node
 
     if is_type(node, NodeType.Model):
-        print_model_result_line(result, schema_name, index, total)
+        printer.print_model_result_line(result, schema_name, index, total)
     elif is_type(node, NodeType.Test):
-        print_test_result_line(result, schema_name, index, total)
+        printer.print_test_result_line(result, schema_name, index, total)
     elif is_type(node, NodeType.Archive):
-        print_archive_result_line(result, index, total)
+        printer.print_archive_result_line(result, index, total)
 
 
-def print_test_result_line(result, schema_name, index, total):
-    model = result.node
-    info = 'PASS'
-    color = Fore.GREEN
+def print_results_line(results, execution_time):
+    stats = {}
 
-    if result.errored:
-        info = "ERROR"
-        color = Fore.RED
+    for result in results:
+        t = result.node.get('resource_type')
 
-    elif result.status > 0:
-        info = 'FAIL {}'.format(result.status)
-        color = Fore.RED
+        if result.node.get('resource_type') == NodeType.Model:
+            t = '{} {}'.format(get_materialization(result.node), t)
 
-        result.fail = True
-    elif result.status == 0:
-        info = 'PASS'
-        color = Fore.GREEN
+        stats[t] = stats.get(t, 0) + 1
 
-    else:
-        raise RuntimeError("unexpected status: {}".format(result.status))
+    stat_line = ", ".join(
+        ["{} {}s".format(ct, t) for t, ct in stats.items()])
 
-    info_text = color + info + Style.RESET_ALL
-
-
-    print_fancy_output_line(
-        "{info} {name}".format(
-            info=info_text,
-            name=model.get('name')),
-        info_text,
-        index,
-        total,
-        result.execution_time)
-
-
-def print_archive_result_line(result, index, total):
-    model = result.node
-    info = 'OK archived'
-
-    if result.errored:
-        info = 'ERROR archiving'
-
-    cfg = model.get('config', {})
-
-    print_fancy_output_line(
-        "{info} {source_schema}.{source_table} --> "
-        "{target_schema}.{target_table}".format(info=info, **cfg),
-        result.status,
-        index,
-        total,
-        result.execution_time)
+    printer.print_timestamped_line("")
+    printer.print_timestamped_line(
+        "Finished running {stat_line} in {execution_time:0.2f}s."
+        .format(stat_line=stat_line, execution_time=execution_time))
 
 
 def execute_test(profile, test):
@@ -214,45 +99,6 @@ def execute_test(profile, test):
             .format(name=test.name, num_cols=len(row)))
 
     return row[0]
-
-
-def print_model_result_line(result, schema_name, index, total):
-    model = result.node
-    info = 'OK created'
-
-    if result.errored:
-        info = 'ERROR creating'
-
-    print_fancy_output_line(
-        "{info} {model_type} model {schema}.{relation}".format(
-            info=info,
-            model_type=get_materialization(model),
-            schema=schema_name,
-            relation=model.get('name')),
-        result.status,
-        index,
-        total,
-        result.execution_time)
-
-
-def print_results_line(results, execution_time):
-    stats = {}
-
-    for result in results:
-        t = result.node.get('resource_type')
-
-        if result.node.get('resource_type') == NodeType.Model:
-            t = '{} {}'.format(get_materialization(result.node), t)
-
-        stats[t] = stats.get(t, 0) + 1
-
-    stat_line = ", ".join(
-        ["{} {}s".format(ct, t) for t, ct in stats.items()])
-
-    print_timestamped_line("")
-    print_timestamped_line(
-        "Finished running {stat_line} in {execution_time:0.2f}s."
-        .format(stat_line=stat_line, execution_time=execution_time))
 
 
 def execute_model(profile, model, existing):
@@ -558,12 +404,11 @@ class RunManager(object):
                 error=str(e),
                 status='ERROR')
 
-        except (RuntimeError,
-                dbt.exceptions.ProgrammingException,
-                psycopg2.ProgrammingError,
-                psycopg2.InternalError) as e:
-            error = "Error executing {filepath}\n{error}".format(
-                filepath=node.get('build_path'), error=str(e).strip())
+        except (RuntimeError, dbt.exceptions.ProgrammingException, psycopg2.ProgrammingError, psycopg2.InternalError) as e: # noqa
+            prefix = dbt.printer.red("Error executing {filepath}\n".format(
+                filepath=node.get('build_path')))
+            error = "{prefix}{error}".format(prefix=prefix, error=str(e).strip())
+
             status = "ERROR"
             logger.debug(error)
             if type(e) == psycopg2.InternalError and \
@@ -574,11 +419,13 @@ class RunManager(object):
                     status="SKIP")
 
         except dbt.exceptions.InternalException as e:
-            error = ("Internal error executing {filepath}\n\n{error}"
+            prefix= 'Internal error executing {filepath}'.format(
+                         filepath=node.get('build_path'))
+            error = ("{prefix}\n\n{error}"
                      "\n\nThis is an error in dbt. Please try again. If "
                      "the error persists, open an issue at "
                      "https://github.com/fishtown-analytics/dbt").format(
-                         filepath=node.get('build_path'),
+                         prefix=dbt.printer.red(prefix),
                          error=str(e).strip())
             status = "ERROR"
 
@@ -687,7 +534,7 @@ class RunManager(object):
         pool = ThreadPool(num_threads)
 
         if should_execute:
-            print_counts(flat_nodes)
+            printer.print_counts(flat_nodes)
 
         start_time = time.time()
 
@@ -702,7 +549,7 @@ class RunManager(object):
         for node_list in node_dependency_list:
             for i, node in enumerate([node for node in node_list
                                       if node.get('skip')]):
-                print_skip_line(node, schema_name, node.get('name'),
+                printer.print_skip_line(node, schema_name, node.get('name'),
                                 get_idx(node), num_nodes)
 
                 node_result = RunModelResult(node, skip=True)
