+++ dbt/task/run.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 from dbt.logger import GLOBAL_LOGGER as logger
 from dbt.runner import RunManager
-import dbt.utils
+import dbt.printer
 
 
 class RunTask:
@@ -17,4 +17,4 @@ class RunTask:
 
         results = runner.run_models(self.args.models, self.args.exclude)
 
-        logger.info(dbt.utils.get_run_status_line(results))
+        logger.info(dbt.printer.get_run_status_line(results))
