+++ dbt/task/test.py
@@ -1,6 +1,6 @@
 from dbt.runner import RunManager
 from dbt.logger import GLOBAL_LOGGER as logger  # noqa
-import dbt.utils
+import dbt.printer
 
 
 class TestTask:
@@ -36,6 +36,6 @@ class TestTask:
         else:
             raise RuntimeError("unexpected")
 
-        logger.info(dbt.utils.get_run_status_line(results))
+        logger.info(dbt.printer.get_run_status_line(results))
 
         return results
